# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Catalogs API Documentation",
        "description": "This is the API documentation for my Laravel application using Swagger/OpenAPI 3.0.",
        "contact": {
            "name": "API Support Team",
            "email": "ptryshaf@gmail.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "",
            "description": "Local development server"
        }
    ],
    "paths": {
...
        "/api/book": {
            "get": {
                "tags": [
                    "Books"
                ],
                "summary": "Get list of book",
                "operationId": "10a8ac3792d996df188aa7aa6f71375c",
                "parameters": [
                    {
                        "name": "search",
                        "in": "query",
                        "description": "Search keyword for book name",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "John"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of book retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "List book successfully"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "title": {
                                                        "type": "string",
                                                        "example": "John Doe"
                                                    },
                                                    "desc": {
                                                        "type": "string",
                                                        "example": "Lorem ipsum dolor sit amet"
                                                    },
                                                    "year_publish": {
                                                        "type": "string",
                                                        "example": "2023-01-01T00:00:00.000000Z"
                                                    },
                                                    "author_id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "publisher_id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "created_at": {
                                                        "type": "string",
                                                        "example": "2023-01-01T00:00:00.000000Z"
                                                    },
                                                    "updated_at": {
                                                        "type": "string",
                                                        "example": "2023-01-01T00:00:00.000000Z"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Books"
                ],
                "summary": "Create a new book",
                "operationId": "a84dc8b3acfa740ffbf6d58086838ab4",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "title",
                                    "desc",
                                    "year_publish",
                                    "author_id",
                                    "publisher_id"
                                ],
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "example": "John Doe"
                                    },
                                    "desc": {
                                        "type": "string",
                                        "example": "Lorem ipsum dolor sit amet"
                                    },
                                    "year_publish": {
                                        "type": "string",
                                        "example": "2023"
                                    },
                                    "author_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "publisher_id": {
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Book created successfully"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                }
            }
        },
        "/api/book/{id}": {
            "get": {
                "tags": [
                    "Books"
                ],
                "summary": "Get detail of a book by ID",
                "operationId": "3b56ae94e9e5f47646e9e758feaed117",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Book ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Book retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "title": {
                                            "type": "string",
                                            "example": "John Doe"
                                        },
                                        "desc": {
                                            "type": "string",
                                            "example": "Lorem ipsum dolor sit amet"
                                        },
                                        "year_publish": {
                                            "type": "string",
                                            "example": "2023"
                                        },
                                        "author_id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "publisher_id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "created_at": {
                                            "type": "string",
                                            "example": "2023-01-01T00:00:00.000000Z"
                                        },
                                        "updated_at": {
                                            "type": "string",
                                            "example": "2023-01-01T00:00:00.000000Z"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "book not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Books"
                ],
                "summary": "Update book",
                "operationId": "8eeb714219a6f11f87d9b74b2c4c7c43",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Book ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "title",
                                    "desc",
                                    "year_publish",
                                    "author_id",
                                    "publisher_id"
                                ],
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "example": "John Doe"
                                    },
                                    "desc": {
                                        "type": "string",
                                        "example": "Lorem ipsum dolor sit amet"
                                    },
                                    "year_publish": {
                                        "type": "string",
                                        "example": "2023"
                                    },
                                    "author_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "publisher_id": {
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "book updated successfully"
                    },
                    "404": {
                        "description": "book not found"
                    },
                    "422": {
                        "description": "Validation failed"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Books"
                ],
                "summary": "Delete a book by ID",
                "operationId": "8196aa437803ac6352e237fb12351d5a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "book ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "book deleted successfully"
                    },
                    "404": {
                        "description": "book not found"
                    }
                }
            }
        },
...
    },
...
}